requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
aiosmtplib==4.0.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
click==8.2.1
dnspython==2.7.0
et_xmlfile==2.0.0
fastapi==0.115.14
frozenlist==1.7.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
multidict==6.6.3
openpyxl==3.1.5
packaging==25.0
pluggy==1.6.0
prometheus_client==0.22.1
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.0.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
yarl==1.20.1
